Modify to store authorities

+++ section_09/spring-sec-section-09/src/main/java/com/ttdat/eazybank/repository/LoanRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 @Repository
 public interface LoanRepository extends CrudRepository<Loans, Long> {
-	List<Loans> findByCustomerIdOrderByStartDtDesc(int customerId);
+	List<Loans> findByCustomerIdOrderByStartDtDesc(long customerId);
 }
